// lib/links.enum.ts
export enum Link {
    VACATIONS_2024 = "https://www.youtube.com/watch?v=yyYTRnRwYn4",
    DIGITAL_MUSEUM = "https://www.youtube.com/watch?v=yyYTRnRwYn4",
    PAYMENT_REPORT = "https://docs.google.com/forms/d/e/1FAIpQLSfu_ANnbXX6HiYkW-S3pYx1LQQC99TSxd1xo_T8G3TmrILiNA/viewform",
    INTERVIEW_REQUEST = "https://docs.google.com/forms/d/e/1FAIpQLSfuhXww4ANlA0j8kMAaFB4b_N0FuJvzjn57WQP4UEw_fvnCYg/viewform",
    INSTAGRAM = "https://www.instagram.com/uecolegiomodestabor/",
    YOUTUBE_CHANNEL = "https://www.youtube.com/@UEColegioModestaBor",
    YOUTUBE_FEATURED = "https://www.youtube.com/watch?v=6QUq8u5BZ7A",
}