// lib/imgs.enum.ts
export enum Imgs {
    LOGO = "/images/emblema-modesta-bor.png",
    MODESTA_BOR_INSCRIPTIONS = "/images/inscription-modesta-bor.png",
    EDUCATIVE_PROJECT = "/images/education-modesta-bor.png",
    MODESTA_BOR_ADVISORY = "/images/advisory-modesta-bor.png",
    PAYMENT_REPORT = "/images/payment-modesta-bor.png",
    MODESTA_BOR = "/images/modesta-bor.png",
    MISSION = "/images/mission-modesta-bor.png",
    VISION = "/images/vision-modesta-bor.png",
    OBJECTIVES = "/images/objectives-modesta-bor.jpg",
    VALUES = "/images/values-modesta-bor.jpg",
    SERVICES_INTRO = "/images/services-intro-modesta-bor.jpg",
    SERVICES_MUSIC = "/images/services-music-modesta-bor.png",
    SERVICES_SPORTS = "/images/services-sports-modesta-bor.png",
    SERVICES_ENGLISH = "/images/services-english-modesta-bor.png",
    SERVICES_SUPPORT = "/images/services-support-modesta-bor.png",
    CONVIVIALITY = "/images/conviviality-modesta-bor.jpg",
}